Setting up the proper inheritance hierarchy.
